// src/app/component/model/car.interface.ts
/**
 * @author giri
 * @email giriramdsk@gmail.com
 * @create date 2020-03-27 16:04:21
 * @modify date 2020-03-27 16:04:21
 * @desc Interface for componend
 */


export class Car {
    make?: string;
    model?: string;
    year?: number[];
    checked?: boolean;
    trim?: string;
    engine?: string;
    selected?: boolean = false;
  
  }
  